requirements.txt
Django==5.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-extensions==3.2.3
python-dotenv==1.0.0
mysqlclient==2.2.0
email-validator==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
django-filter==23.5
openpyxl==3.1.2
reportlab==4.0.7 